// src/pages/main-page/types.ts
export type TScreenData = {
  accentColor: string;
  metaTitle: string;
  projectsTitle: string;
  projectsSubtitle: string;
  projectsMetaLeft?: string;
  projectsMetaRight?: string;
  navLeftLabel: string;
  navRightLabel: string;
  navLeftLink: string;
  navRightLink: string;
  contactEmail: string;
  linkedInUrl: string;
};